--- include/moss/ecs/camera.hpp
@@ -1,20 +1,0 @@
-/*
-ecs/camera.hpp - app / parentless
-
-Does not implement standard ecs components/systems.
-Default created by core. Custom can be created by
-user.
-
-*/
-
-#pragma once
-
-#include <moss/includes.hpp>
-#include <moss/ecs/components.hpp>
-#include <moss/ecs/systems.hpp>
-#include <raylib.h>
-
-namespace moss {
-
-
-} // moss
