+++ include/moss/ecs/ecs.hpp
@@ -1,0 +1,14 @@
+/*
+ecs/ecs.hpp - parentless
+
+Wrapper file for including all ecs
+based files.
+
+*/
+
+
+#pragma once
+
+#include <moss/ecs/components.hpp>
+#include <moss/ecs/serialization.hpp>
+#include <moss/ecs/systems.hpp>
