+++ include/moss/meta/libs.hpp
@@ -1,0 +1,17 @@
+/*
+moss/libs.hpp - parentless
+
+Main hub for including libaries that
+is in usage for moss.
+
+*/
+
+
+#pragma once
+
+#include <entt/entt.hpp>
+#include <glm/glm.hpp>
+#include <GLFW/glfw3.h>
+
+#include <nlohmann/json.hpp>
+using nlohmann::json;
