+++ include/moss/moss.hpp
@@ -1,0 +1,14 @@
+/*
+moss/moss.hpp - parentless
+
+Main hub for including moss as user.
+
+*/
+
+#pragma once
+
+
+#include <moss/ecs/ecs.hpp>
+#include <moss/meta/defines.hpp>
+#include <moss/meta/versions.hpp>
+#include <moss/utils/utils.hpp>
