+++ include/moss/utils/utils.hpp
@@ -1,0 +1,11 @@
+/*
+utils/utils.hpp - parentless
+
+Includes the core utils. Wrapper file for
+the whole util namespace.
+
+*/
+
+#pragma once
+
+#include <moss/utils/math.hpp>
